document-search-using-agent-engine-agent-framework/backend/requirements.txt
gunicorn==20.1.0; python_version > '3.0'
gunicorn==19.10.0; python_version < '3.0'
uvicorn~=0.17.0
fastapi~=0.111.1
fastapi-utilities==0.2.0
cloudpickle
h5py==3.8.0
grpcio==1.66.2
urllib3==1.26.16
typing-inspect==0.9.0
typing_extensions==4.12.2
pydantic
requests==2.31.0

langchain==0.3.2
langchain-community==0.3.1
langchain-core==0.3.9
langchain-google-vertexai==2.0.4

google-cloud-aiplatform[adk,agent-engines]==1.86.0
google-cloud-bigquery==3.26.0
google-cloud-tasks==2.16.5
google-cloud-logging==3.11.2
google-cloud-speech==2.27.0
google-cloud-discoveryengine==0.13.4

